requirements.txt
streamlit
pandas
PyPDF2
openai
<<<<<<< HEAD
xlsxwriterpython-docx
openai
=======
xlsxwriter
python-docx
>>>>>>> a06ba87f1f10134ab3c57fb44e2961897e6278f9
